pyproject.toml
[project]
name = "pdfbaker"
version = "0.9.2"
description = "SVG Jinja templates + YAML config = PDF documents"
authors = [
    { name = "Danny W. Adair", email = "danny.adair@unfold.nz" }
]
dependencies = [
    "cairosvg",
    "click",
    "jinja2",
    "pydantic",
    "pypdf",
    "ruamel.yaml",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
pdfbaker = "pdfbaker.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
dev-dependencies = [
    # --- Testing
    "pytest",
    "pytest-cov",

    # --- Security
    "pip-audit",
    "bandit",

    # --- Release
    "python-semantic-release"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pdfbaker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=pdfbaker --cov-report=term-missing"

[tool.coverage.run]
source = ["pdfbaker"]

[tool.pylint.main]
py-version = "3.11"
init-hook = "import sys; sys.path.insert(0, 'src')"

[tool.pylint.messages_control]
disable = ["W0511"]  # Disable TODO/FIXME warnings

[tool.pylint.reports]
msg-template = "{path}:{line}: [{msg_id}({symbol}), {obj}] {msg}"
output-format = "colorized"
reports = "no"
score = "no"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["src/pdfbaker/__init__.py:__version__"]
branch = "main"
build_command = """
uv lock
git add uv.lock
uv build
"""
tag_format = "{version}"
commit_version_number = true
commit_message = "RELEASE: {version}"

[tool.semantic_release.changelog]
template_dir = ".templates"

[tool.semantic_release.changelog_sections]
feat = "Features"
fix = "Bug Fixes"
docs = "Documentation"
style = "Style"
refactor = "Refactor"
perf = "Performance"
test = "Tests"
build = "Build"
ci = "CI"
chore = "Chores"
revert = "Reverts"

[tool.semantic_release.changelog_exclude]
types = ["style", "test", "build", "ci", "chore"]

[tool.semantic_release.changelog_section_headers]
feat = "### Features"
fix = "### Bug Fixes"
docs = "### Documentation"
refactor = "### Refactor"
perf = "### Performance"
revert = "### Reverts"
